clang: Make libclang_frontend depend on libbcc
Libclang uses functions from libbcc, so it does depend on it.
When defining new symbols in bpf_common.cc, linker errors appear
when the same symbols are used in libclang. This is because of
incorrect linker dependency. This patch fixes the issue by making
sure the dependency is correctly tracked.

Signed-off-by: Joel Fernandes <joelaf@google.com>