Redirect syscount.py's reads of "comm" files to BPFd for remote targets
This queries BPFd for the "comm" file of a given pid when a remote
target is configured. This fixes the issue with syscount.py, which was
reading "comm" files locally instead of on the remote target.

This bug caused "syscount.py -P" to show "[unknown]" when running for a
remote target, since it was erroneously looking up the target's
process pids on the local host.

Signed-off-by: Jazel Canseco <jcanseco@google.com>