Fix bug where failed file reads on remote are not detected by libremote
BPFd responds with the message "Open failed, ignoring" when it fails to
open a file it is requested to read out to stdout. Before this patch,
libremote did not account for the fact that BPFd may respond with this
error message, and erroneously let it pass through as a successful
response.

Signed-off-by: Jazel Canseco <jcanseco@google.com>